+++ .vscode/settings.json
@@ -19,8 +19,10 @@
     "autodocs",
     "autoprefixer",
     "autosize",
+    "backmerge",
     "backticked",
     "beitzah",
+    "bethclark",
     "bugsnag",
     "categorisation",
     "Categorised",
@@ -28,6 +30,7 @@
     "cloudinary",
     "clsx",
     "codegen",
+    "COLOR",
     "compat",
     "contrib",
     "cuid",
@@ -40,6 +43,7 @@
     "Docgen",
     "dockerized",
     "dopplerhq",
+    "dotenv",
     "EASS",
     "EHRC",
     "estree",
@@ -106,6 +110,7 @@
     "ponyfill",
     "popover",
     "portabletext",
+    "postcss",
     "posthog",
     "postpack",
     "posttest",
@@ -117,6 +122,7 @@
     "psql",
     "pusherapp",
     "ratelimit",
+    "refs",
     "Regen",
     "remeda",
     "Rerank",
@@ -128,6 +134,7 @@
     "sslmode",
     "SUBJ",
     "superjson",
+    "svgs",
     "tailwindcss",
     "tanstack",
     "testid",
@@ -144,6 +151,7 @@
     "uidotdev",
     "unjudged",
     "unsets",
+    "unshallow",
     "unsummarised",
     "untruncate",
     "untruncated",
@@ -155,6 +163,7 @@
     "valign",
     "vars",
     "vectorstores",
+    "vercel",
     "WCAG",
     "webvtt",
     "zadd",
@@ -194,5 +203,6 @@
   "yaml.schemas": {
     "file:///c%3A/Users/Simon/.vscode/extensions/atlassian.atlascode-2.8.6/resources/schemas/pipelines-schema.json": "bitbucket-pipelines.yml",
     "https://www.artillery.io/schema.json": []
-  }
+  },
+  "typescript.preferences.preferTypeOnlyAutoImports": true
 }
