+++ apps/nextjs/src/app/api/chat/errorHandling.ts
@@ -4,7 +4,7 @@ import {
   ErrorDocument,
 } from "@oakai/aila/src/protocol/jsonPatchProtocol";
 import { handleHeliconeError } from "@oakai/aila/src/utils/moderation/moderationErrorHandling";
-import { UserBannedError } from "@oakai/core/src/models/safetyViolations";
+import { UserBannedError } from "@oakai/core/src/models/userBannedError";
 import { TracingSpan } from "@oakai/core/src/tracing/serverTracing";
 import { RateLimitExceededError } from "@oakai/core/src/utils/rateLimiting/userBasedRateLimiter";
 import { PrismaClientWithAccelerate } from "@oakai/db";
