+++ apps/nextjs/src/app/api/chat/webActionsPlugin.ts
@@ -1,11 +1,11 @@
-import { AilaThreatDetectionError } from "@oakai/aila";
 import { AilaPlugin } from "@oakai/aila/src/core/plugins";
+import { AilaThreatDetectionError } from "@oakai/aila/src/features/threatDetection";
 import { handleHeliconeError } from "@oakai/aila/src/utils/moderation/moderationErrorHandling";
 import {
   SafetyViolations as defaultSafetyViolations,
   inngest,
 } from "@oakai/core";
-import { UserBannedError } from "@oakai/core/src/models/safetyViolations";
+import { UserBannedError } from "@oakai/core/src/models/userBannedError";
 import { PrismaClientWithAccelerate } from "@oakai/db";
 import { aiLogger } from "@oakai/logger";
 import { waitUntil } from "@vercel/functions";
