+++ packages/aila/package.json
@@ -44,6 +44,7 @@
     "@pollyjs/persister-fs": "^6.0.6",
     "@types/jest": "^29.5.12",
     "eslint": "^8.56.0",
+    "eslint-plugin-turbo": "^2.2.3",
     "jest": "^29.7.0",
     "setup-polly-jest": "^0.11.0",
     "ts-jest": "^29.1.4",
