+++ packages/aila/src/core/Aila.liveWithOpenAI.test.ts
@@ -1,7 +1,7 @@
-import type { AilaInitializationOptions } from ".";
-import { Aila } from ".";
 import { MockCategoriser } from "../features/categorisation/categorisers/MockCategoriser";
+import { Aila } from "./Aila";
 import { checkLastMessage, expectPatch, expectText } from "./Aila.testHelpers";
+import type { AilaInitializationOptions } from "./types";
 
 const runInCI = process.env.CI === "true";
 const runManually = process.env.RUN_LLM_TESTS === "true";
