+++ packages/aila/src/core/chat/AilaChat.ts
@@ -5,17 +5,13 @@ import {
 } from "@oakai/core/src/utils/subjects";
 import invariant from "tiny-invariant";
 
-import type { AilaChatService, AilaServices } from "../..";
-import { AilaError } from "../..";
 import { DEFAULT_MODEL, DEFAULT_TEMPERATURE } from "../../constants";
-import type {
-  AilaGenerationStatus} from "../../features/generation";
-import {
-  AilaGeneration
-} from "../../features/generation";
+import type { AilaChatService } from "../../core/AilaServices";
+import type { AilaServices } from "../../core/AilaServices";
+import { AilaGeneration } from "../../features/generation/AilaGeneration";
+import type { AilaGenerationStatus } from "../../features/generation/types";
 import { generateMessageId } from "../../helpers/chat/generateMessageId";
-import type {
-  JsonPatchDocumentOptional} from "../../protocol/jsonPatchProtocol";
+import type { JsonPatchDocumentOptional } from "../../protocol/jsonPatchProtocol";
 import {
   LLMMessageSchema,
   parseMessageParts,
@@ -24,6 +20,7 @@ import type {
   AilaPersistedChat,
   AilaRagRelevantLesson,
 } from "../../protocol/schema";
+import { AilaError } from "../AilaError";
 import type { LLMService } from "../llm/LLMService";
 import { OpenAIService } from "../llm/OpenAIService";
 import type { AilaPromptBuilder } from "../prompt/AilaPromptBuilder";
@@ -279,7 +276,7 @@ export class AilaChat implements AilaChatService {
       invariant(responseText, "Response text not set");
       await this._generation.complete({ status, responseText });
     }
-    this._generation.persist(status);
+    await this._generation.persist(status);
   }
 
   private async persistChat() {
