+++ packages/aila/src/core/chat/PatchEnqueuer.test.ts
@@ -14,6 +14,7 @@ describe("PatchEnqueuer", () => {
 
     await patchEnqueuer.enqueuePatch(path, value);
 
+    // eslint-disable-next-line @typescript-eslint/unbound-method
     expect(controller.enqueue).toHaveBeenCalled();
     const expectedPatch = `\n␞\n${JSON.stringify({
       type: "patch",
@@ -21,6 +22,7 @@ describe("PatchEnqueuer", () => {
       value: { op: "add", path, value },
       status: "complete",
     })}\n␞\n`;
+    // eslint-disable-next-line @typescript-eslint/unbound-method
     expect(controller.enqueue).toHaveBeenCalledWith(expectedPatch);
   });
