+++ packages/aila/src/features/americanisms/AilaAmericanisms.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/triple-slash-reference
 /// <reference path="./american-british-english-translator.d.ts" />
-import { textify } from "@oakai/core/src/models/lessonPlans";
+import { textify } from "@oakai/core/src/utils/textify";
 import translator from "american-british-english-translator";
 
 import type { AilaAmericanismsFeature } from ".";
