+++ packages/aila/src/features/generation/AilaGeneration.ts
@@ -4,12 +4,12 @@ import {
   generateAilaPromptVersionVariantSlug,
 } from "@oakai/core/src/prompts/lesson-assistant/variants";
 import type { Prompt } from "@oakai/db";
-import { prisma } from "@oakai/db";
+import { prisma } from "@oakai/db/client";
 import { aiLogger } from "@oakai/logger";
 import { kv } from "@vercel/kv";
 import { getEncoding } from "js-tiktoken";
 
-import type { AilaServices } from "../../core";
+import type { AilaServices } from "../../core/AilaServices";
 import type { AilaChat } from "../../core/chat";
 import type { AilaGenerationStatus } from "./types";
