+++ packages/aila/src/features/generation/index.test.ts
@@ -1,8 +1,8 @@
-import { AilaGeneration } from ".";
-import type { AilaInitializationOptions } from "../..";
-import { Aila } from "../..";
+import { Aila } from "../../core/Aila";
 import type { Message } from "../../core/chat";
 import { AilaChat } from "../../core/chat";
+import type { AilaInitializationOptions } from "../../core/types";
+import { AilaGeneration } from "./AilaGeneration";
 
 const ailaArgs: AilaInitializationOptions = {
   plugins: [],
@@ -21,7 +21,9 @@ describe("calculateTokenUsage", () => {
     ];
 
     const mockEncoding = {
-      encode: jest.fn().mockImplementation((text) => text.split(" ").length),
+      encode: jest
+        .fn()
+        .mockImplementation((text: string) => text.split(" ").length),
     };
     jest.mock("js-tiktoken", () => ({
       getEncoding: () => mockEncoding,
@@ -57,7 +59,9 @@ describe("calculateTokenUsage", () => {
     ];
 
     const mockEncoding = {
-      encode: jest.fn().mockImplementation((text) => text.split(" ").length),
+      encode: jest
+        .fn()
+        .mockImplementation((text: string) => text.split(" ").length),
     };
     jest.mock("js-tiktoken", () => ({
       getEncoding: () => mockEncoding,
