+++ packages/aila/src/features/moderation/getSessionModerations.ts
@@ -1,6 +1,6 @@
-import { Moderations } from "@oakai/core";
-import type { Moderation} from "@oakai/db";
-import { prisma } from "@oakai/db";
+import { Moderations } from "@oakai/core/src/models/moderations";
+import { prisma } from "@oakai/db/client";
+import type { Moderation } from "@prisma/client";
 
 export async function getSessionModerations(
   appSessionId: string,
