+++ packages/aila/src/features/moderation/index.test.ts
@@ -1,12 +1,13 @@
 // Generated by CodiumAI
-import type { Moderations } from "@oakai/core";
+import type { Moderations } from "@oakai/core/src/models/moderations";
 import type { ModerationResult } from "@oakai/core/src/utils/ailaModeration/moderationSchema";
 import type { PrismaClientWithAccelerate } from "@oakai/db";
 
 import { AilaModeration } from ".";
-import { Aila } from "../..";
-import type { AilaChatInitializationOptions, Message } from "../../core";
+import { Aila } from "../../core/Aila";
+import type { Message } from "../../core/chat";
 import type { AilaPlugin } from "../../core/plugins";
+import type { AilaChatInitializationOptions } from "../../core/types";
 import type { LooseLessonPlan } from "../../protocol/schema";
 import type { AilaModerator } from "./moderators";
 import { MockModerator } from "./moderators/MockModerator";
@@ -182,6 +183,7 @@ describe("AilaModeration", () => {
         messages,
       };
       const moderations = {
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-return
         create: jest.fn((mod) => ({ id: "ABC", ...mod })),
       } as unknown as Moderations;
       const mockPlugin = {
@@ -211,6 +213,7 @@ describe("AilaModeration", () => {
         categories: moderationResult.categories,
         id: "ABC",
       });
+      // eslint-disable-next-line @typescript-eslint/unbound-method
       expect(mockPlugin.onToxicModeration).toHaveBeenCalledWith(
         expect.objectContaining({
           categories: ["t/encouragement-illegal-activity"],
