+++ packages/aila/src/features/moderation/moderators/OpenAiModerator.ts
@@ -1,10 +1,7 @@
 import { createOpenAIClient } from "@oakai/core/src/llm/openai";
 import { moderationPrompt } from "@oakai/core/src/utils/ailaModeration/moderationPrompt";
-import type {
-  ModerationResult} from "@oakai/core/src/utils/ailaModeration/moderationSchema";
-import {
-  moderationResponseSchema,
-} from "@oakai/core/src/utils/ailaModeration/moderationSchema";
+import type { ModerationResult } from "@oakai/core/src/utils/ailaModeration/moderationSchema";
+import { moderationResponseSchema } from "@oakai/core/src/utils/ailaModeration/moderationSchema";
 import { aiLogger } from "@oakai/logger";
 import type OpenAI from "openai";
 import type {
@@ -18,7 +15,7 @@ import {
   DEFAULT_MODERATION_MODEL,
   DEFAULT_MODERATION_TEMPERATURE,
 } from "../../../constants";
-import type { AilaServices } from "../../../core";
+import type { AilaServices } from "../../../core/AilaServices";
 
 const log = aiLogger("aila:moderation");
 
@@ -141,14 +138,14 @@ export class OpenAiModerator extends AilaModerator {
         message: "No message.content in moderation response from OpenAI",
         data: { moderationResponse },
       });
-      throw new AilaModerationError(`Failed to get moderation response`);
+      throw new AilaModerationError("Failed to get moderation response");
     }
     if (!response.data) {
       this._aila?.errorReporter?.addBreadcrumb({
         message: "Invalid moderation response",
         data: { moderationResponse },
       });
-      throw new AilaModerationError(`No moderation response`);
+      throw new AilaModerationError("No moderation response");
     }
 
     const { categories, justification, scores } = response.data;
