+++ packages/aila/src/features/persistence/adaptors/kv/index.ts
@@ -1,7 +1,7 @@
 import { kv } from "@vercel/kv";
 
 import { AilaPersistence } from "../..";
-import type { AilaChatService, AilaServices } from "../../../../core";
+import type { AilaChatService, AilaServices } from "../../../../core/AilaServices";
 
 export class AilaKVPersistence extends AilaPersistence {
   constructor({ chat, aila }: { chat: AilaChatService; aila: AilaServices }) {
