+++ packages/aila/src/features/threatDetection/index.ts
@@ -1,2 +1,3 @@
 export { AilaThreatDetection } from "./AilaThreatDetection";
 export { AilaThreatDetector } from "./detectors/AilaThreatDetector";
+export { AilaThreatDetectionError } from "./types";
