+++ packages/aila/src/features/types.ts
@@ -3,7 +3,7 @@ import type { AilaPluginContext } from "../core/plugins";
 import type { ModerationDocument } from "../protocol/jsonPatchProtocol";
 import type { AilaPersistedChat, LooseLessonPlan } from "../protocol/schema";
 import type { AilaErrorBreadcrumb, AilaErrorSeverity } from "./errorReporting/types";
-import type { AilaGeneration } from "./generation";
+import type { AilaGeneration } from "./generation/AilaGeneration";
 import type { AilaThreatDetector } from "./threatDetection";
 
 export interface AilaModerationFeature {
