+++ packages/aila/src/protocol/jsonPatchProtocol.ts
@@ -1,19 +1,13 @@
 import { moderationCategoriesSchema } from "@oakai/core/src/utils/ailaModeration/moderationSchema";
 import { aiLogger } from "@oakai/logger";
 import * as Sentry from "@sentry/nextjs";
-import type {
-  Operation} from "fast-json-patch";
-import {
-  applyPatch,
-  deepClone,
-  JsonPatchError,
-} from "fast-json-patch";
+import type { Operation } from "fast-json-patch";
+import { applyPatch, deepClone, JsonPatchError } from "fast-json-patch";
 import untruncateJson from "untruncate-json";
 import { z } from "zod";
 import zodToJsonSchema from "zod-to-json-schema";
 
-import type {
-  LooseLessonPlan} from "./schema";
+import type { LooseLessonPlan } from "./schema";
 import {
   BasedOnOptionalSchema,
   BasedOnSchema,
