+++ packages/core/src/llm/openai.ts
@@ -1,13 +1,5 @@
 import type { OpenAIProvider } from "@ai-sdk/openai";
 import { createOpenAI } from "@ai-sdk/openai";
-import { ChatOpenAI as LangchainChatOpenAI } from "langchain/chat_models/openai";
-import type { BaseLLMParams } from "langchain/llms/base";
-import type {
-  AzureOpenAIInput,
-  OpenAIInput} from "langchain/llms/openai";
-import {
-  OpenAI as OpenAILangchain,
-} from "langchain/llms/openai";
 import type { ClientOptions } from "openai";
 import OpenAI from "openai";
 
@@ -57,54 +49,8 @@ function createOpenAIModerationsClient() {
   return openAi.moderations;
 }
 
-function createOpenAILangchainClient({
-  app,
-  fields = {},
-}: {
-  app: string;
-  fields?: Partial<OpenAIInput> &
-    Partial<AzureOpenAIInput> &
-    BaseLLMParams & {
-      configuration?: ClientOptions;
-    };
-}) {
-  const defaultHeaders = heliconeHeaders({ app });
-  return new OpenAILangchain({
-    ...fields,
-    configuration: {
-      apiKey: process.env.OPENAI_API_KEY,
-      baseURL: process.env.HELICONE_EU_HOST,
-      defaultHeaders,
-    },
-  });
-}
-
-function createOpenAILangchainChatClient({
-  app,
-  fields = {},
-}: {
-  app: string;
-  fields?: Partial<OpenAIInput> &
-    Partial<AzureOpenAIInput> &
-    BaseLLMParams & {
-      configuration?: ClientOptions;
-    };
-}) {
-  const defaultHeaders = heliconeHeaders({ app });
-  return new LangchainChatOpenAI({
-    ...fields,
-    configuration: {
-      apiKey: process.env.OPENAI_API_KEY,
-      baseURL: process.env.HELICONE_EU_HOST,
-      defaultHeaders,
-    },
-  });
-}
-
 export {
   createOpenAIClient,
   createVercelOpenAIClient,
   createOpenAIModerationsClient,
-  createOpenAILangchainClient,
-  createOpenAILangchainChatClient,
 };
