+++ packages/core/src/models/apps.ts
@@ -1,5 +1,5 @@
-import type { PrismaClientWithAccelerate, Prompt } from "@oakai/db";
-import { Prisma } from "@oakai/db";
+import type { Prompt, PrismaClientWithAccelerate } from "@oakai/db";
+import { Prisma } from "@prisma/client";
 import type { App } from "@prisma/client";
 
 import { Prompts } from "./prompts";
