+++ packages/core/src/models/prompts.ts
@@ -7,7 +7,7 @@ import type { BaseMessage} from "langchain/schema";
 import { SystemMessage } from "langchain/schema";
 import untruncateJson from "untruncate-json";
 
-import { createOpenAILangchainChatClient } from "../llm/openai";
+import { createOpenAILangchainChatClient } from "../llm/langchain";
 
 type CompletionMeta = {
   timeTaken: number;
@@ -133,7 +133,7 @@ export class Prompts {
             }
           },
           handleChainError: (err) => {
-            this.logger.error(err, `LLM chain error`);
+            this.logger.error(err, "LLM chain error");
           },
         },
       ],
@@ -162,7 +162,7 @@ export class Prompts {
     }
 
     if ("errorMessage" in result && typeof result.errorMessage === "string") {
-      this.logger.info(`llm refused to fulfil prompt: %s`, result.errorMessage);
+      this.logger.info("llm refused to fulfil prompt: %s", result.errorMessage);
       throw new LLMRefusalError(result.errorMessage, meta);
     }
