+++ packages/core/src/models/userBannedError.ts
@@ -1,0 +1,5 @@
+export class UserBannedError extends Error {
+  constructor(userId: string) {
+    super(`User banned: ${userId}`);
+  }
+}
