+++ packages/core/src/utils/getCaptionsFromFile.ts
@@ -1,6 +1,6 @@
 import { Storage } from "@google-cloud/storage";
 import { aiLogger } from "@oakai/logger";
-import type { Cue} from "webvtt-parser";
+import type { Cue } from "webvtt-parser";
 import { WebVTTParser } from "webvtt-parser";
 
 const log = aiLogger("transcripts");
